requirements.txt
asgiref==3.4.1
certifi==2022.12.7
charset-normalizer==2.0.12
Django==3.2.17
django-cors-headers==3.10.1
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
PyMySQL==1.0.2
pytz==2022.7.1
requests==2.27.1
sqlparse==0.4.3
typing_extensions==4.1.1
urllib3==1.26.14
